Fix BT4 autocast quantile crash

+++ src/chess_nn_playground/models/architecture/bt4_mixers/legal_move_graph_delta.py
@@ -72,7 +72,7 @@ class LegalMoveGraphDeltaMixer(nn.Module):
         # (B, R, 64, 64) relation-wise scores.
         scores = torch.einsum("birk,bjrk->brij", s, d)
         q = 1.0 - min(max(self._density, 1.0 / _SQUARES), 1.0)
-        thresh = torch.quantile(scores, q, dim=-1, keepdim=True)
+        thresh = torch.quantile(scores.float(), q, dim=-1, keepdim=True)
         return (scores >= thresh).to(dtype=tokens.dtype)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
